@@ -12,10 +12,7 @@ use Symfony\Component\Messenger\MessageBusInterface as Bus;
 
 abstract class AbstractBusType extends AbstractType
 {
-    /**
-     * @var Bus
-     */
-    private $bus;
+    private Bus $bus;
 
     public function __construct(Bus $bus)
     {
